Refactor: moving /api/search handler to handlers_custom_api
